compat: add getpeereid.c and move compat out of control.c
